requirements.txt
streamlit
numpy
matplotlib
scipy
plotly.express